♻️ Revise hosted ComposeView handling to avoid composition issues
